% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc.R
\name{is_spc}
\alias{is_spc}
\title{Is spc}
\usage{
is_spc(x)
}
\arguments{
\item{x}{object to test}
}
\value{
logical inidicating whether `x` is an spc object or not
}
\description{
Convenience function that checks if an object is an spc object
}
