% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_i.R
\name{is_spc_i}
\alias{is_spc_i}
\title{Is spc i}
\usage{
is_spc_i(x)
}
\arguments{
\item{x}{object to test}
}
\value{
logical inidicating whether `x` is an spc i object or not
}
\description{
Convenience function that checks if an object is an spc i object
}
