% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHSRspc-package.R
\docType{package}
\name{NHSRspc-package}
\alias{NHSRspc}
\alias{NHSRspc-package}
\title{NHSRspc: NHS-R SPC}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

This package provides a modern approach to Statistical Process Control in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tomjemmett/NHSRspc}
  \item Report bugs at \url{https://github.com/tomjemmett/NHSRspc/issues}
}

}
\author{
\strong{Maintainer}: Tom Jemmett \email{thomas.jemmett@nhs.net} (\href{https://orcid.org/0000-0002-6943-2990}{ORCID})

Other contributors:
\itemize{
  \item NHS-R community \email{nhs.rcommunity@nhs.net} [copyright holder]
}

}
\keyword{internal}
