% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_i.R
\name{spc_i}
\alias{spc_i}
\title{SPC I}
\usage{
spc_i(
  x = seq_along(y),
  y = numeric(),
  centre = NULL,
  limits = 3,
  exclude = rep(FALSE, length(y)),
  freeze = NULL
)
}
\arguments{
\item{x}{numeric vector to order y by, defaults to an incrementing sequence (assuming the provided order of y is
correct)}

\item{y}{numeric vector that we want to calculate control limits for}

\item{centre}{the centre of y, defaults to the mean of y}

\item{limits}{the multiplier used to calculate limits, defaults to +/- 3}

\item{exclude}{a logical vector the same length of y, indicating points to exclude from control limit calculations}

\item{freeze}{a number indicating after how many points to freeze control limit calculations after}
}
\value{
an spc_i object
}
\description{
Calculates limits for an SPC-I chart
}
\examples{
set.seed(123)
y <- rnorm(10)
spc_i(y = y)

x <- sample(1:10)
spc_i(x, y)

spc_i(x, y, centre = 3)

spc_i(x, y, exclude = sample(c(TRUE, FALSE),
                             size = length(y),
                             prob = c(0.3, 0.7),
                             replace = TRUE))
}
